templates/django/djangoRest/requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.0
defusedxml==0.7.0rc1
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-bootstrap4==2.2.0
django-celery-results==1.2.1
dnspython==1.16.0
eventlet==0.29.1
filelock==3.0.12
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
instaloader==4.5.3
kombu==4.6.11
more-itertools==8.5.0
oauthlib==3.1.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyotp==2.4.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
robin-stocks==1.5.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.33
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.1.0
